requirements.txt
fastapi
uvicorn[standard]
websockets
python-multipart
python-dotenv
requests
opencv-python
pillow
pydantic
python-jose[cryptography]
passlib[bcrypt]